Candidates page is now partially functional, add by excel functionality still not implemented

+++ server/src/controllers/templates.controller.js
@@ -4,46 +4,47 @@ import fs from "fs";
 import path from "path";
 
 export const createTemplate = async (req, res) => {
-   const categoryName = req.params.categoryName;
+   const categoryUri = req.params.categoryId;
    const { name, description, authorId } = req.fields;
    const tempFile = req.files.example;
 
-   const trimmedName = name.toLowerCase().replace(/\s/g, "-");
-   const examplePath = path.join(process.cwd(), "uploads", "templates", trimmedName + ".pdf");
-
-   const templateExists = await Template.findOne({ name: trimmedName });
-   if (templateExists) return res.json({ warning: "Este documento ya existe" });
-
    const templateCreated = await new Template({
-      name: trimmedName,
+      name: name,
       description: description,
-      category: categoryName,
-      examplePath: examplePath,
+      category: categoryUri,
       createdBy: authorId,
-   }).save();
+   });
+   const examplePath = path.join(process.cwd(), "uploads", "templates", templateCreated._id + ".pdf");
+   templateCreated.examplePath = examplePath;
+   templateCreated.save();
 
    fs.writeFileSync(examplePath, fs.readFileSync(tempFile.path));
    fs.unlinkSync(tempFile.path);
 
    // Updates category by adding ID of new template
-   await Category.findOneAndUpdate(
-      { name: categoryName },
+   const something = await Category.findOneAndUpdate(
+      { uri: categoryUri },
       {
          $push: {
             templates: templateCreated._id
          }
+      },
+      {
+         new: true
       }
    );
 
-   //file: fs.readFileSync(examplePath)
+   console.log(something);
+
    res.json(templateCreated);
 }
 
 
 export const getTemplates = async (req, res) => {
-   const categoryName = req.params.categoryName;
-   const templatesFound = await Template.find({ category: categoryName });
+   const categoryUri = req.params.categoryId;
+   const templatesFound = await Template.find({ category: categoryUri });
 
+   // Creates a new array of templates with the data of the corresponding filepaths
    let templatesComputed = [];
    for (let i = 0; i < templatesFound.length; i++) {
       templatesComputed.push({ ...templatesFound[i]._doc, example: fs.readFileSync(templatesFound[i]._doc.examplePath) });
@@ -53,23 +54,22 @@ export const getTemplates = async (req, res) => {
 }
 
 export const updateTemplateById = async (req, res) => {
-   const categoryName = req.params.categoryName;
    const templateId = req.params.templateId;
    const { name, description, authorId } = req.fields;
    const tempFile = req.files.example;
 
-   const trimmedName = name.toLowerCase().replace(/\s/g, "-");
-   const examplePath = path.join(process.cwd(), "uploads", "templates", trimmedName + ".pdf");
+   const templateExists = await Template.findOne({ _id: templateId });
+   if (!templateExists) {
+      fs.unlinkSync(tempFile.path);
+      return res.json({ warning: "Este documento ya no existe" });
+   }
 
-   const templateExists = await Template.findOne({ name: trimmedName });
-   if (templateExists) return res.json({ warning: "Este documento ya existe" });
-
+   const examplePath = path.join(process.cwd(), "uploads", "templates", templateExists._id + ".pdf");
    const templateUpdated = await Template.findOneAndUpdate(
       { _id: templateId },
       {
-         name: trimmedName,
+         name: name,
          description: description,
-         category: categoryName,
          examplePath: examplePath,
          updatedBy: authorId
       },
@@ -78,7 +78,7 @@ export const updateTemplateById = async (req, res) => {
       }
    )
 
-   // Deletes previous file
+   // Finds and deletes previous file
    fs.unlinkSync(templateExists.examplePath);
 
    // Creates new file
@@ -91,20 +91,22 @@ export const updateTemplateById = async (req, res) => {
 }
 
 export const deleteTemplateById = async (req, res) => {
-   const categoryName = req.params.categoryName;
+   const categoryUri = req.params.categoryId;
    const templateId = req.params.templateId;
 
    const templateExists = await Template.findOne({ _id: templateId });
    if (!templateExists) return res.json({ warning: "Este documento ya no existe" });
 
-   const templateDeleted = await Template.findOneAndDelete({ _id: templateId });
+   // Deletes existing file
+   if (fs.existsSync(templateExists.examplePath)) {
+      fs.unlinkSync(templateExists.examplePath);
+   }
 
-   // Deletes existing file file
-   fs.unlinkSync(templateExists.examplePath);
+   const templateDeleted = await Template.findOneAndDelete({ _id: templateId });
 
    // Updates category by deleting ID of deleted template
    await Category.findOneAndUpdate(
-      { name: categoryName },
+      { uri: categoryUri },
       {
          $pull: {
             templates: templateId
